pyproject.toml
[tool.poetry]
name = "pytels-finances-back"
version = "0.1.0"
description = ""
authors = ["raul-gomes <rsgomes86@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
